% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_patterns_var.R
\name{binary_patterns_var}
\alias{binary_patterns_var}
\title{Combines (usually multiple) binary vectors into a convenient
single vector summarising the pattern}
\usage{
binary_patterns_var(
  data,
  strip_prefix = "",
  mid_sep = ", ",
  last_sep = " & ",
  none_label = "(None)"
)
}
\arguments{
\item{data}{Data frame or tibble containing the binary
vectors}

\item{strip_prefix}{A common prefix to remove from
the start of the variable names}

\item{mid_sep}{Character to separate names, by default a comma}

\item{last_sep}{The last separator to use, by default an ampersand}

\item{none_label}{How to represent the empty combination}
}
\value{
A character vector representing combinations of inputs
}
\description{
This is useful when you have a collection of yes/no variables
representing multiple choices in which more than one option
can be selected.
}
\examples{
test1 <- expand.grid(myvar_1 = c(0,1), 
                     myvar_2 = c(0,1),
                     myvar_3 = c(0,1))
test1$pat <- binary_patterns_var(test1, "myvar_")
test1
}
