% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativity.R
\name{contracted_length}
\alias{contracted_length}
\title{Calculates length contraction}
\usage{
contracted_length(proper_length, speed)
}
\arguments{
\item{proper_length}{proper length (in metres) from the
perspective of the rest frame}

\item{speed}{Speed of moving object in m/s}
}
\value{
Returns the contracted length in metres
}
\description{
Calculates length contraction from the perspective of a
moving object's rest frame
}
