% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enum_contingency_tables.R
\name{enum_contingency_tables}
\alias{enum_contingency_tables}
\title{Enumerate contingency tables with fixed grand total}
\usage{
enum_contingency_tables(nrow, ncol, total)
}
\arguments{
\item{nrow}{Number of rows in each matrix}

\item{ncol}{Number of columns}

\item{total}{Grand total the cells sum to}
}
\value{
A list of nrow by ncol matrices
}
\description{
This function enumerates all contingency tables of a particular
size all summing to a grand total.
}
