% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativity.R
\name{light_speed}
\alias{light_speed}
\title{Returns a multiple (by default, 1) of the speed of light in m/s}
\usage{
light_speed(m = 1)
}
\arguments{
\item{m}{Multiple of the speed of light}
}
\value{
Returns m * the speed of light in m/s
}
\description{
Returns a multiple (by default, 1) of the speed of light in m/s
}
\examples{
light_speed() # the speed of light in m/s
light_speed(0.5) # half the speed of light in m/s
}
