% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativity.R
\name{proper_length}
\alias{proper_length}
\title{Calculates proper length}
\usage{
proper_length(contracted_length, speed)
}
\arguments{
\item{contracted_length}{Length from the perspective of
the moving object}

\item{speed}{Speed of moving object in m/s}
}
\value{
Returns the proper length in metres
}
\description{
Calculates proper length (perspective of moving object's rest frame)
given the moving object's contracted length
}
