% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativity.R
\name{proper_time}
\alias{proper_time}
\title{Calculates proper time}
\usage{
proper_time(dilated_time, speed)
}
\arguments{
\item{dilated_time}{The time on the stationary object. Whatever units
are chosen here will also be the units of the results.}

\item{speed}{Speed of moving object in m/s}
}
\value{
Returns the proper time in units chosen for dilated_time
}
\description{
Calculates proper time (perspective of moving object)
from time as measured on a stationary object
}
