% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativity.R
\name{dilated_time}
\alias{dilated_time}
\title{Calculates dilated time}
\usage{
dilated_time(proper_time, speed)
}
\arguments{
\item{proper_time}{The time on the object. Whatever units
are chosen here will also be the units of the results.}

\item{speed}{Speed of moving object in m/s}
}
\value{
Returns the dilated time in units chosen for proper_time
}
\description{
Calcualted dilated time (perspective of stationary observer)
from proper time experienced on object moving at a particular
speed
}
