% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativity.R
\name{dilated_time}
\alias{dilated_time}
\title{Calculates dilated time}
\usage{
dilated_time(proper_time, speed)
}
\arguments{
\item{proper_time}{The time (in seconds) as experienced on the moving object}

\item{speed}{Speed of moving object in m/s}
}
\value{
Returns the dilated time in seconds
}
\description{
Calculates dilated time (perspective of stationary observer)
from proper time experienced on object moving at a given
speed
}
\examples{
curve(dilated_time(1, light_speed(x)), 0, .99,
      xlab = "Speed (as proprtion of speed of light)",
      ylab = "Dilated time",
      main = "Dilated time when proper time = 1")
}
